pyproject.toml
[tool.poetry]
name = "python-attestation-task"
version = "0.1.0"
description = ""
authors = ["Sprite_Spirit <Sprite.spirit.t.me@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
pillow = "^11.0.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}
django-countries = "^7.6.1"
drf-spectacular = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
